chore(ls): refactor into files
Splits the 4000 LOC ls.rs codebase into three of ~1000-1500 each. Makes the codebase more manageable.
